+++ varfish/static/js/state_machine.js
@@ -220,7 +220,8 @@ function handleEventStateInitial(eventType, event) {
             displayConnectionError();
             break;
           default:
-            console.log("Error during AJAX call: ", jqXHR, textStatus, errorThrown);
+            alert("Error during AJAX call:", textStatus, + errorThrown, "in state", currentState);
+            console.log("Error during AJAX call: ", jqXHR, textStatus, errorThrown, currentState);
             handleEvent(EVENT_ERROR, null);
         }
       }
@@ -264,7 +265,8 @@ function handleEventStateIdle(eventType, event) {
             handleEvent(EVENT_ERROR, null);
             break;
           default:
-            console.log("Error during AJAX call: ", jqXHR, textStatus, errorThrown);
+            alert("Error during AJAX call:", textStatus, + errorThrown, "in state", currentState);
+            console.log("Error during AJAX call: ", jqXHR, textStatus, errorThrown, currentState);
             handleEvent(EVENT_ERROR, null);
         }
       }
@@ -306,7 +308,8 @@ function handleEventStateGetJobId(eventType, event) {
             displayConnectionError();
             break;
           default:
-            console.log("Error during AJAX call: ", jqXHR, textStatus, errorThrown);
+            alert("Error during AJAX call:", textStatus, + errorThrown, "in state", currentState);
+            console.log("Error during AJAX call: ", jqXHR, textStatus, errorThrown, currentState);
             handleEvent(EVENT_ERROR, null);
         }
       }
@@ -396,7 +399,8 @@ function handleEventStateWaitJobResults(eventType, event) {
             displayConnectionError();
             break;
           default:
-            console.log("Error during AJAX call: ", jqXHR, textStatus, errorThrown);
+            alert("Error during AJAX call:", textStatus, + errorThrown, "in state", currentState);
+            console.log("Error during AJAX call: ", jqXHR, textStatus, errorThrown, currentState);
             handleEvent(EVENT_ERROR, null);
         }
       }